Rename `vue` config to `vue2`.
**BREAKING**: Rename `vue` config to `vue2`. Various projects have been
using `vue` when they should be using `vue3`. This will make usage more
explicit.
