+++ templates/vue2.js
@@ -1,5 +1,5 @@
 import digitalbazaar from 'eslint-config-digitalbazaar';
-import digitalbazaarVue from 'eslint-config-digitalbazaar/vue';
+import digitalbazaarVue from 'eslint-config-digitalbazaar/vue2';
 import globals from 'globals';
 
 export default [
